+++ Control/Joint/Effects/Reader.hs
@@ -2,7 +2,7 @@ module Control.Joint.Effects.Reader where
 
 import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
-import Control.Joint.Abilities.Transformer (Transformer (Schema, embed, build, unite), (:>) (T))
+import Control.Joint.Abilities.Transformer (Transformer (embed, build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (TU (TU))
 
@@ -22,8 +22,9 @@ instance Interpreted (Reader e) where
 	type Primary (Reader e) a = (->) e a
 	run (Reader x) = x
 
+type instance Schema (Reader e) = TU ((->) e)
+
 instance Transformer (Reader e) where
-	type Schema (Reader e) = TU ((->) e)
 	embed x = T . TU . const $ x
 	build x = T. TU $ pure <$> run x
 	unite = T . TU