+++ Control/Joint/Effects/Writer.hs
@@ -2,7 +2,7 @@ module Control.Joint.Effects.Writer where
 
 import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
-import Control.Joint.Abilities.Transformer (Transformer (Schema, embed, build, unite), (:>) (T))
+import Control.Joint.Abilities.Transformer (Transformer (embed, build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
 import Control.Joint.Schemes (UT (UT))
 
@@ -24,8 +24,9 @@ instance Interpreted (Writer e) where
 	type Primary (Writer e) a = (e, a)
 	run (Writer x) = x
 
+type instance Schema (Writer e) = UT ((,) e)
+
 instance Monoid e => Transformer (Writer e) where
-	type Schema (Writer e) = UT ((,) e)
 	embed x = T . UT $ (,) mempty <$> x
 	build = T . UT . pure . run
 	unite = T . UT